feat!: support DSPi 1.1.5 V16 wire format
This is a breaking change that drops compatibility with pre-V16
firmware. Key changes:

- Add chunked bulk transfers (0xA2/0xA3) to stay within Linux
  usbfs 4096-byte limit
- Replace hardcoded offset-based bulk access with a dynamic
  field registry
- Bump max channels to 17 (8 inputs + 9 outputs) for RP2350
- Update ClipFlags from uint16 to uint32
- Add 5 new EQ filter types (Notch, AllPass, AllPass1,
  LowShelf1, HighShelf1)
- Generalize I2S RX pin API with pair-aware accessors
- Add I2S input channel count support (0xF3/0xF4)
- Make channel completions dynamic via NumInputChannels
  instead of hardcoded splits

+++ snapshot.go
@@ -44,7 +44,7 @@ type MeterSnapshot struct {
 	Peaks     [maxChannels]Level
 	CPU0      int
 	CPU1      int
-	ClipFlags uint16
+	ClipFlags uint32
 	Channels  int
 	err       error
 }
